[fixed] auto reconstruction of author name by clicking the input
The input field "myusernameedit" is'nt in use, because the author name
is fetched from users' profile. So it is set "readonly" in the custom
pad.js file. But by clicking the input the full name will be filled in
and replace the user defined field value, if chosen. This fix converts
the input into a span.

In addition a small translation improvement is provided.